Error log deletion minor.
git-svn-id: https://plugins.svn.wordpress.org/post-pay-counter/tags/2.22@876073 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ classes/ppc_ajax_functions_class.php
@@ -199,8 +199,10 @@ class PPC_ajax_functions {
         global $ppc_global_settings;
         self::ppc_check_ajax_referer( 'ppc_clear_error_log' );
         
-        if( ! delete_option( $ppc_global_settings['option_errors'] ) ) {
-            die( __( 'Error: could not clear error log.', 'post-pay-counter' ) );
+        if( get_option( $ppc_global_settings['option_errors'] ) ) {
+            if( ! delete_option( $ppc_global_settings['option_errors'] ) ) {
+                die( __( 'Error: could not clear error log.', 'post-pay-counter' ) );
+            }
         }
         
         die( 'ok' );
